feat: useModelInfoStore separated into two files
