+++ src/services/global/LiveModel/ExportLM.ts
@@ -1,10 +1,10 @@
 import { Message } from '../../../components/lit-components/message-wrapper';
 import config from '../../../config';
 import useControlStore from '../../../stores/LiveModel/ControlStore/useControlStore';
+import useModelInfoStore from '../../../stores/LiveModel/ModelInfoStore/useModelInfoStore';
 import useRegulationsStore from '../../../stores/LiveModel/RegulationsStore/useRegulationsStore';
 import useUpdateFunctionsStore from '../../../stores/LiveModel/UpdateFunctionsStore/useUpdateFunctionsStore';
 import useLoadedModelStore from '../../../stores/LiveModel/useLoadedModelStore';
-import useModelInfoStore from '../../../stores/LiveModel/useModelInfoStore';
 import useVariablesStore from '../../../stores/LiveModel/VariablesStore/useVariablesStore';
 import type {
   ControlInfo,
